services/api/requirements.txt
Flask
flask_pymongo
flask-restx
Flask-Cors
botocore
boto3
python-dotenv
json-logging